Submission commit of HW1.

+++ src/convert_knots.h
@@ -1,5 +1,5 @@
 double knots_to_miles_per_minute(int knots) {
-    double feetPerHour = 6076 * knots;
+    double feetPerHour = 6076 * double(knots);
     double milePerHour = feetPerHour / 5280;
     double milePerMinute = milePerHour / 60;
     return milePerMinute;
